third_party/blink/web_tests/fast/forms/select/option-defaultselected-expected.txt

Tests for HTMLOptionElement::defaultSelected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS option1.defaultSelected is false
PASS option1.defaultSelected = true; option1.hasAttribute("selected") is true
PASS option1.selected is true
PASS option1.selected = false; option1.defaultSelected is true
PASS option1.defaultSelected = false; option1.hasAttribute("selected") is false
PASS option1.setAttribute("selected", "no"); option1.defaultSelected is true
PASS option1.removeAttribute("selected"); option1.defaultSelected is false
PASS selectionMap(select1) is "100"
PASS select1[2].defaultSelected = true; selectionMap(select1) is "001"
PASS select1[1].defaultSelected = true; selectionMap(select1) is "010"
PASS select1[1].defaultSelected = false; selectionMap(select1) is "100"
PASS select1[2].selected = true; selectionMap(select1) is "001"
PASS successfullyParsed is true

TEST COMPLETE
